oops
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@6089 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sound/MidiFile.cpp
@@ -1503,7 +1503,7 @@ MidiFile::convertToMidi(Composition &comp)
 				soundingDuration * m_timingDivision /
 				crotchetDuration;
 
-			    int pitch = 60;
+			    long pitch = 60;
 			    (*el)->get<Int>(BaseProperties::PITCH, pitch);
 			    pitch += (*segment)->getTranspose();
